event: add typed signed event state
- add typed validated event envelopes and tag primitives
- preserve canonical wire JSON in signed event state
- add optional direct Schnorr signature verification
- align codec consumers with private envelope access

+++ crates/event/src/follow.rs
@@ -20,7 +20,7 @@ pub struct RadrootsFollow {
 )]
 #[derive(Clone, Debug)]
 pub struct RadrootsFollowProfile {
-    pub published_at: u32,
+    pub published_at: u64,
     pub public_key: String,
     pub relay_url: Option<String>,
     pub contact_name: Option<String>,
